Swagger added

+++ src/main/java/com/ankush/gotstock/dto/StockHoldingDTO.java
@@ -6,6 +6,8 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Pattern;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class StockHoldingDTO {
 
@@ -22,11 +24,11 @@ public class StockHoldingDTO {
 
     @NotNull(message = "Purchase price is mandatory")
     @Min(value = 0, message = "Purchase price cannot be negative")
-    private Double purchasePrice;
+    private BigDecimal purchasePrice;
 
-    private Double currentPrice;
+    private BigDecimal currentPrice;
 
-    private Double gainLoss;
+    private BigDecimal gainLoss;
 
 
 }
